Try adding a wrap to the output state of cartpole
I'm still puzzling through why the cartpole NN isn't swinging up. It seems like when I train it to have a target position of "down" it wants to hold an angle a little off from the downwards fixed point, so there still might be a bug going on. I thought perhaps it was because the cartpole state wraps from 0 to 2pi, so I added an explicit wrap system to the diagram, but that didn't seem to change it.