Make LQR work on the revised cartpole urdf
The reference frames for some of the angles were flipped, so I changed the starting and goal state to reflect that. I also am using a `PlanarSceneGraphVisualizer` instead of MeshCat, since it's easier to run and view without needing the browser open.

I also added a plot of input force over time.